Don't parse GBT transactions; leave as hex. (LTC MWEB fix)
This commit also removes support for share versions < 35, which I did because
I was a bit too lazy to think about some of the pre-34 code paths (which used
transactions more heavily). This support could be re-added in a later commit
if there is demand for it.

This commit also disables the block publication over p2p because I was lazy,
and that should be fixed soon.